improved show.html.erb & tooltips
Show.html.erb now displays image details under the image, rather than in a tooltip, which improves its usability for mobile users.
Index.html.erb tooltips have been modified so that they are properly escaped and now appear as multiline tooltips.

While Index tooltips are not visible to mobile users, I decided that if an brick's details couldn't be visible as a tooltip, it would be better that the details not be visible at all, especially since the details can be found by clicking on the image. (In the future, I might create a flash notice alerting new users to "Tap an image to view more details")